requirements.txt
fastapi
passlib
pydantic
PyJWT
python-dotenv
python_jose
sqlmodel
uvicorn
psycopg2
pydantic[email]
python-multipart
pymilvus
openai